Create New House
Marin and Nellie buy a house not far from Sofia. Nellie loves flowers so much that she convinces you to write a program that will calculate how much it will cost them to plant a certain number of flowers and whether the available budget will be enough for them. Different flowers have different prices.

Price per piece in BGN: 
Rose - 5.00; Dahlia - 3.80; Tulip - 2.80; Narcissus - 3.00; Gladiola - 2.50.

The following discounts are available:
• If Nellie buys more than 80 Roses - 10% discount on the final price
• If Nellie buys more than 90 Dahlias - 15% discount from the final price
• If Nellie buys more than 80 Tulips - 15% discount from the final price
• If Nellie buys less than 120 Narcissus - the price increases by 15%
• If Nellie Buys less than 80 Gladioli - the price increases by 20%

3 lines are read from the console:
• Flower type - text with options - "Roses", "Dahlias", "Tulips", "Narcissus", "Gladiolus"
• Number of flowers - an integer in the range [10…1000]
• Budget - an integer in the range [50…2500]

Print to the console on one line:
• If their budget is sufficient - "Hey, you have a great garden with {number of flowers} {type of flowers} and {remaining amount} leva left."
• If their budget is NOT enough - "Not enough money, you need {necessary amount} leva more."
Amount to be formatted to the second decimal place.