Create Fishing Boat
Tony and friends loved to go fishing, they are so passionate about fishing that they decided to go fishing by boat. The price for renting the vessel depends on the season and the number of fishermen.

The price depends on the season:
• The price for renting the ship in the spring is BGN 3,000.
• The price for renting the ship in summer and autumn is BGN 4,200.
• The price for renting the ship in winter is BGN 2,600.

Depending on their number, the group benefits from a discount:
• If the group is up to 6 people inclusive - 10% discount.
• If the group is from 7 to 11 people inclusive - 15% discount.
• If the group is 12 or more - 25% discount.

Fishermen enjoy an additional 5% discount if they are an even number, unless it is autumn - then they do not have an additional discount.

Write a program to calculate whether the fishermen will collect enough money.

Login
Exactly three lines are read from the console.
• Group budget – an integer in the range [1…8000]
• Season - text: "Spring", "Summer", "Autumn", "Winter"
• Number of fishermen – an integer in the range [4…18]

Exit
To print one line to the console:
• If the budget is sufficient:
"Yes! You have {the remaining money} leva left."
• If the budget IS NOT sufficient:
"Not enough money! You need {amount that does not reach} leva."

Amounts must be formatted to two decimal places.