Create Journey
Strange, but most people plan their vacation early. A young programmer has a certain budget and free time in a given season. Write a program that accepts the budget and the season as input, and outputs where the programmer will rest and how much he will spend.
The budget determines the destination and the season determines how much of the budget will be spent. If it's summer, he'll rest at a campsite, and in winter at a hotel. If he is in Europe, regardless of the season, he will rest in a hotel. Each campsite or hotel, according to the destination, has its own price that corresponds to a certain percentage of the budget:

• At BGN 100 or less - somewhere in Bulgaria
o Summer – 30% of the budget
o Winter – 70% of the budget

• At BGN 1,000. or less - somewhere in the Balkans
o Summer – 40% of the budget
o Winter – 80% of the budget

• For more than BGN 1,000. – somewhere in Europe
o When traveling around Europe, regardless of the season, will spend 90% of the budget.

Login
The input is read from the console and consists of two lines entered by the user:
• First line – Budget, a real number in the interval [10.00...5000.00].
• Second line – One of the two possible seasons: “summer” or “winter”.

Exit
Two lines should be printed on the console.
• First line – "Somewhere in [destination]" between "Bulgaria", "Balkans" and "Europe"
• Second row – "{Type of vacation} – {Amount spent}"

 Rest can be between "Camp" and "Hotel"
 The amount must be rounded to the second decimal place.