Create Operation between Numbers
Write a program that reads two integers (N1 and N2) and an operator to perform a given mathematical operation on them. The possible operations are: Addition(+), Subtraction(-), Multiplication(*), Division(/) and Modular Division(%). Addition, subtraction and multiplication on the console should print the result and whether it is even or odd. In ordinary division - the result. In the case of modular division – the remainder. It should be noted that the divisor can be equal to 0(zero), and zero is not divisible. In this case, a special message should be printed.

Login: 3 lines entered by the user are read from the console

• N1 – an integer in the interval [0...40 000]
• N2 – an integer in the interval [0...40 000]
• Operator - one character between: "+", "-", "*", "/", "%"

Exit: To print one line to the console

• If the operation is addition, subtraction or multiplication:
 "{N1} {operator} {N2} = {result} - {even/odd}"
• If the operation is division:  "{N1} / {N2} = {result}" - the result is formatted to the second decimal place
• If the operation is modular division:  "{N1} % {N2} = {remainder}"
• In case of division by 0(zero):  "Cannot divide {N1} by zero"