Create On Time For The Exam
A student must go to an exam at a certain time. He comes to the exam hall at a given arrival time. The student is considered to have arrived on time if he arrived at the time of the exam or up to half an hour before. If he arrived more than 30 minutes earlier, he was late. If he came after the exam time, he is late. Write a program that reads an exam time and an arrival time and prints whether the student was on time, whether he was early or late, and by how many hours or minutes he was early or late.

Login: 4 integers (one per line) entered by the user are read from the console
• The first line contains the exam time - an integer from 0 to 23.
• The second line contains the minute of the exam - an integer from 0 to 59.
• The third line contains the arrival time - an integer from 0 to 23.
• The fourth line contains the arrival minute - an integer from 0 to 59.

Exit: On the first line, print
• “Late” if the student arrives later than the exam time.
• "On time", if the student arrives exactly at the time of the exam or up to 30 minutes earlier.
• “Early” if the student arrives more than 30 minutes before the exam time.