Create Ski Trip
Atanas decides to spend his vacation in Bansko and go skiing. Before he goes, however, he needs to book a hotel and calculate how much the stay will cost him. The following room types are available with the following stay rates:

 "room for one person" – BGN 18.00 per night
 "apartment" – BGN 25.00 per night
 "presidential suite" – BGN 35.00 per night

According to the number of days he will stay in the hotel (example: 11 days = 10 nights) and the type of room he will choose, he can use a different discount.

The reductions are as follows:
   
less than 10 days - room for one person does not use a discount
between 10 and 15 days - room for one person does not use a discount
more than 15 days room - for one person does not use a discount

less than 10 days - apartment 30% of the final price
between 10 and 15 days - 35% of the final price
more than 15 days room - 50% of the final price

less than 10 days - presidential suite 10% of the final price
between 10 and 15 days - 15% of the final price 
more than 15 days room -  20% of the final price

After the stay, Atanas' assessment of the hotel services can be positive or negative. If his assessment is positive, Atanas adds 25% of it to the already discounted price. In case of a negative evaluation, 10% of the price is deducted.

The input is read from the console and consists of three lines:
• First row - days of stay - an integer in the interval [0...365]
• Second row - room type - "single room", "apartment" or "presidential suite"
• Third row - evaluation - "positive" or "negative"

Exit: One line should be printed to the console
• The price of his hotel stay, formatted to the second decimal place.